refactor(lotto): purchaseAmount의 형 변경 (Long -> Int)

+++ src/main/kotlin/lotto/utils/PurchaseAmountCalculator.kt
@@ -3,7 +3,7 @@ package lotto.utils
 import lotto.constants.Lotto
 
 object PurchaseAmountCalculator {
-  fun calculatePurchaseAmount(purchaseMoney: Long): Long {
-    return purchaseMoney / Lotto.LOTTO_PRICE
+  fun calculatePurchaseAmount(purchaseMoney: Long): Int {
+    return (purchaseMoney / Lotto.LOTTO_PRICE).toInt()
   }
 }
