+++ src/main/kotlin/lotto/view/LottoIssueView.kt
@@ -4,7 +4,7 @@ import lotto.constants.GuideMessage.SHOW_ISSUED_AMOUNT
 
 object LottoIssueView {
 
-  fun printIssuedAmountMessage(purchaseAmount: Long) {
+  fun printIssuedAmountMessage(purchaseAmount: Int) {
     println(SHOW_ISSUED_AMOUNT.format(purchaseAmount))
   }
